+++ src/main/java/de/unistuttgart/ims/coref/annotator/Constants.java
@@ -149,6 +149,8 @@ public class Constants {
 			"nl", "ru", "gmh" };
 
 	public static final int UI_MAX_STRING_WIDTH_IN_TREE = 50;
+	public static final int UI_MAX_STRING_WIDTH_IN_MENU = 30;
+	public static final int UI_MAX_STRING_WIDTH_IN_STATUSBAR = 30;
 
 	public static final String URL_LATEST_RELEASE_API = "https://api.github.com/repos/nilsreiter/CorefAnnotator/releases/latest";